pyproject.toml
[tool.poetry]
name = "google-sheets"
version = "0.1.0"
description = ""
authors = ["Daniel <matsubus@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
black = "^22.3.0"
isort = "^5.10.1"
python-dotenv = "^0.20.0"
pydantic = "^1.9.1"
SQLAlchemy = "^1.4.37"
psycopg2-binary = "^2.9.3"
psycopg2 = "^2.9.3"
gspread = "^5.4.0"
oauth2client = "^4.1.3"
google-api-python-client = "^2.51.0"
requests = "^2.28.0"
APScheduler = "^3.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
